combined HttpRequestAction and BaseHttpRequestAction into one
